+++ CHANGELOG.md
@@ -10,10 +10,22 @@ If you like this project and find it useful, please consider giving it a star on
 
 ## [3.0.3] - 2025-05-18
 
+### New plugins
+
+[Dyson robot](https://github.com/thoukydides/matterbridge-dyson-robot)
+
+A Matterbridge plugin that connects Dyson robot vacuums and air treatment devices.
+to the Matter smart home ecosystem via their local MQTT APIs.
+
+[Aeg robot](https://github.com/thoukydides/matterbridge-aeg-robot)
+
+AEG RX 9 / Electrolux Pure i9 robot vacuum plugin for Matterbridge.
+
 ### Added
 
 - [virtual] Added virtual devices configuration mode in the Matterbridge Settings: 'Disabled', 'Light', 'Outlet', 'Switch', 'Mounted_switch'. Switch is not supported by Alexa. Mounted Switch is not supported by Apple.
 - [deviceTypes] Added evse, waterHeater, solarPower, batteryStorage and heatPump device type.
+- [waterHeater] Added WaterHeater class to create a Water Heater Device Type in one line of code (thanks https://github.com/lboue).
 
 ### Changed
 
@@ -21,6 +33,7 @@ If you like this project and find it useful, please consider giving it a star on
 - [export]: Removed long deprecated Matter exports from matterbridge. Use matterbridge/matter.
 - [matterbridge]: Refactored initialize() and cleanup() methods.
 - [matterbridge]: Updated -help informations.
+- [rvc]: Added the parameters in the RoboticVacuumCleaner class constructor.
 
 ### Fixed